% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential_main.R
\name{credentialing}
\alias{credentialing}
\title{#' Credentialing main function}
\usage{
credentialing(peaktable1, peaktable2, ppm, rtwin, rtcom, ratio1 = 1/1,
  ratio2 = 1/2, ratio_tol = 0.1, ratios_tol = 0.1, cd = 13.00335 - 12,
  charges = 1:2, mpc = c(12, 120), maxnmer = 4)
}
\arguments{
\item{peaktable1}{the features from the first ratio dataset.}

\item{peaktable2}{the features from the second ratio dataset.}

\item{ppm}{mass error tolerance for isotopologue searching and credentialed features searching.}

\item{rtwin}{retention time window for 1st round credentialing}

\item{rtcom}{retention time window for 2nd round credentialing}

\item{ratio1}{ratio of designated 12C/13C ratio in the first peaktable, by default set to 1/1}

\item{ratio2}{ratio of designated 12C/13C ratio in the first peaktable, by default set to 1/2}

\item{ratio_tol}{a factor that controls the ratio range to pass the 1st round credentialing, The default value is 0.1.}

\item{ratios_tol}{a factor that controls the ratio range to pass the 2nd round credentialing, The default value is 0.1.}

\item{cd}{unit mass difference among isotopologues. by defalut is 13C-12C = 13.00335 - 12.}

\item{charges}{charge states of the ions to be considered when searching isotope pairs. The default value is 1:2.}

\item{mpc}{A range of mass per carbon to be considered when searching possible isotope pairs and identifying U12C and U13C isotope peaks. The default value is c(12,120)}

\item{maxnmer}{maximum number of ion aggregation to be considered, The default value is 4.}
}
\value{
A list of the credentialed features before and after 1st and 2nd round filtering by the function.
}
\description{
This is the main function for credentialing.
}
\keyword{credentialing}
