% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsGroupExtract.R
\name{xsGroupExtract}
\alias{xsGroupExtract}
\title{Extract peak-table from an xcmsSet object}
\usage{
xsGroupExtract(xs, intchoice = "into", sampling = 1)
}
\arguments{
\item{xs}{a xcmsSet object}

\item{intchoice}{the intensity options provided by xcms. Options: "into"(default) -raw peak area, "intb" -baseline corrected peak area, "maxo" - peak height}

\item{sampling}{the way to sampling intensity in each feature group. If 1, sampling mean intensities of all peaks. If 2, sampling the maximum intensity among all peaks.}
}
\value{
a list includes two data.tables. featuretable is the detailed featuretable. credTable is the simplified format credentialing accepts in the next step.
}
\description{
Extract peak-table from an xcmsSet object
}
